refactor: move the library-write actions into src/api/actions.rs
main.rs 2195 -> 1841. These seven functions are the write half of the Web API
and were the last HTTP code left in the binary root.

The rest of src/api/ reads - fetch, GET, paging. This file writes: save and
unsave, follow and unfollow, playlist adds. build_action_menu is the odd one
out, building the menu model that describes those writes rather than
performing one, and it lives here because the writes it describes do.

Pure move. No logic changed - only pub(crate) on the seven functions, the use
statements, and rustfmt rewrapping the spawn_action_menu and spawn_action
signatures that pub(crate) pushed past 100 columns. No visibility changes were
needed on ActionKind, ActionItem, ActionMenu or LibItem; a private item at the
crate root is already visible to descendant modules.

139 tests still pass; fmt, clippy and test clean with --all-features under
-D warnings.
